[nss-host] Increase sleep time for defunct all responses
In a huge traffic load like 4K, 1 second sleep time
is not sufficient to receive all the ACK/NACK responses
from the NSS fw for the destroy commands before removing
the module completely. While we are removing ECM with the
rmmod command, it stops and exits several modules. To avoid
races between these processes, we should make sure we received
all the responses from NSS fw. 5 seconds sleep time is sufficient
for this.

Change-Id: Ic7cc52333e1439904a59a3c9e7838052d98992b4
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>