[nss-host] Updating the RPS script for Spliy wifi (RDP 454)
Requirement:
In split wifi (RDP 454)
1) rxfill_0, txcmpl_0 & pci0_* related
interrupts should be mapped to core2
2) rxfill_1, txcmpl_1 & pci2_* related
interrupts should be mapped to core1

Fix:
Added board check in qca-nss-ppe-ds.init file
to identify the board and execute the IRQ
affinity for the respective board.
Also sets the IRQ affinity for
ppe2tcl and reo2ppe ring interrupts for RDP 433.

Change-Id: I5337033bda81ee82b6a05ce0cdf5272a0ced438c
Signed-off-by: Nanda Krishnan <quic_nandkris@quicinc.com>